Split unsigned.ts into module with per-version files
